Fix Wallet Increase 3 Digit seperate

+++ DNSLab.Web/Interfaces/Repositories/IPaymentRepository.cs
@@ -6,7 +6,7 @@ namespace DNSLab.Web.Interfaces.Repositories
     public interface IPaymentRepository
     {
         Task<bool> Verify(long trackId);
-        Task<string?> RequestPaymentUrl(int amount);
+        Task<string?> RequestPaymentUrl(long amount);
         Task<PagedResult<PaymentDTO>?> GetPayments(int page = 1, int pageSize = 10);
         Task<PagedResult<PaymentDTO>?> GetAllPayments(int page = 1, int pageSize = 10);
     }
